Emit correct keys to the index
Since keys can be broken up among nodes special care needs to be taken to
create correct index structure. This means that only the first and new key
per node (with respect to eor flag) is allowed to act as the first key.
This also correctly sets the last key if the key is to be broken up.